+++ rtl/sys_gen/tst_sram/cmoda7/sys_tst_sram_c7.vhd
@@ -1,6 +1,6 @@
--- $Id: sys_tst_sram_c7.vhd 984 2018-01-02 20:56:27Z mueller $
+-- $Id: sys_tst_sram_c7.vhd 1086 2018-12-16 18:29:55Z mueller $
 --
--- Copyright 2017- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+-- Copyright 2017-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 --
 -- This program is free software; you may redistribute and/or modify it under
 -- the terms of the GNU General Public License as published by the Free
@@ -15,8 +15,7 @@
 -- Module Name:    sys_tst_sram_c7 - syn
 -- Description:    test of cmoda7 sram and its controller
 --
--- Dependencies:   vlib/xlib/s7_cmt_sfs
---                 vlib/genlib/clkdivce
+-- Dependencies:   bplib/bpgen/s7_cmt_1ce1ce
 --                 bplib/bpgen/bp_rs232_2line_iob
 --                 vlib/rlink/rlink_sp2c
 --                 tst_sram
@@ -29,7 +28,7 @@
 -- Test bench:     tb/tb_tst_sram_c7
 --
 -- Target Devices: generic
--- Tool versions:  viv 2017.1; ghdl 0.34
+-- Tool versions:  viv 2017.1-2018.2; ghdl 0.34
 --
 -- Synthesized (viv):
 -- Date         Rev  viv    Target       flop  lutl  lutm  bram  slic
@@ -37,6 +36,7 @@
 --
 -- Revision History: 
 -- Date         Rev Version  Comment
+-- 2018-12-16  1086   1.1    use s7_cmt_1ce1ce
 -- 2017-06-11   914   1.0    Initial version
 -- 2017-06-11   912   0.5    First draft (derived from sys_tst_sram_n4)
 ------------------------------------------------------------------------------
@@ -135,56 +135,34 @@ architecture syn of sys_tst_sram_c7 is
 
 begin
 
-  GEN_CLKSYS : s7_cmt_sfs               -- clock generator system ------------
+  GEN_CLKALL : s7_cmt_1ce1ce            -- clock generator system ------------
     generic map (
-      VCO_DIVIDE     => sys_conf_clksys_vcodivide,
-      VCO_MULTIPLY   => sys_conf_clksys_vcomultiply,
-      OUT_DIVIDE     => sys_conf_clksys_outdivide,
       CLKIN_PERIOD   => 83.3,
       CLKIN_JITTER   => 0.01,
       STARTUP_WAIT   => false,
-      GEN_TYPE       => sys_conf_clksys_gentype)
+      CLK0_VCODIV    => sys_conf_clksys_vcodivide,
+      CLK0_VCOMUL    => sys_conf_clksys_vcomultiply,
+      CLK0_OUTDIV    => sys_conf_clksys_outdivide,
+      CLK0_GENTYPE   => sys_conf_clksys_gentype,
+      CLK0_CDUWIDTH  => 7,
+      CLK0_USECDIV   => sys_conf_clksys_mhz,
+      CLK0_MSECDIV   => 1000,      
+      CLK1_VCODIV    => sys_conf_clkser_vcodivide,
+      CLK1_VCOMUL    => sys_conf_clkser_vcomultiply,
+      CLK1_OUTDIV    => sys_conf_clkser_outdivide,
+      CLK1_GENTYPE   => sys_conf_clkser_gentype,
+      CLK1_CDUWIDTH  => 7,
+      CLK1_USECDIV   => sys_conf_clkser_mhz,
+      CLK1_MSECDIV   => 1000)
     port map (
-      CLKIN   => I_CLK12,
-      CLKFX   => CLK,
-      LOCKED  => open
-    );
-  
-  CLKDIV_CLK : clkdivce                 -- usec/msec clock divider system ----
-    generic map (
-      CDUWIDTH => 7,                    -- good for up to 127 MHz !
-      USECDIV  => sys_conf_clksys_mhz,
-      MSECDIV  => 1000)
-    port map (
-      CLK     => CLK,
-      CE_USEC => CE_USEC,
-      CE_MSEC => CE_MSEC
-    );
-
-  GEN_CLKSER : s7_cmt_sfs               -- clock generator serport------------
-    generic map (
-      VCO_DIVIDE     => sys_conf_clkser_vcodivide,
-      VCO_MULTIPLY   => sys_conf_clkser_vcomultiply,
-      OUT_DIVIDE     => sys_conf_clkser_outdivide,
-      CLKIN_PERIOD   => 83.3,
-      CLKIN_JITTER   => 0.01,
-      STARTUP_WAIT   => false,
-      GEN_TYPE       => sys_conf_clkser_gentype)
-    port map (
-      CLKIN   => I_CLK12,
-      CLKFX   => CLKS,
-      LOCKED  => open
-    );
-
-  CLKDIV_CLKS : clkdivce                -- usec/msec clock divider serport ---
-    generic map (
-      CDUWIDTH => 7,
-      USECDIV  => sys_conf_clkser_mhz,
-      MSECDIV  => 1000)
-    port map (
-      CLK     => CLKS,
-      CE_USEC => open,
-      CE_MSEC => CES_MSEC
+      CLKIN     => I_CLK12,
+      CLK0      => CLK,
+      CE0_USEC  => CE_USEC,
+      CE0_MSEC  => CE_MSEC,
+      CLK1      => CLKS,
+      CE1_USEC  => open,
+      CE1_MSEC  => CES_MSEC,
+      LOCKED    => open
     );
 
   IOB_RS232 : bp_rs232_2line_iob